вынес ActionCreator и ActionType в отдельные файлы
