+++ src/action-type.js
@@ -1,0 +1,4 @@
+export const ActionType = {
+  CHANGE_CITY: `CHANGE_CITY`,
+  CHANGE_OFFERS: `CHANGE_OFFERS`
+};
