+++ src/reducer.js
@@ -1,3 +1,4 @@
+import {ActionType} from "./action-type";
 import {offers} from "./mocks/offers";
 
 const activeCity = `Amsterdam`;
@@ -5,27 +6,6 @@ const activeCity = `Amsterdam`;
 const initialState = {
   activeCity,
   offers
-};
-
-const ActionType = {
-  CHANGE_CITY: `CHANGE_CITY`,
-  CHANGE_OFFERS: `CHANGE_OFFERS`
-};
-
-export const ActionCreator = {
-  changeActiveCity: (changedCity) => {
-    return {
-      type: ActionType.CHANGE_CITY,
-      payload: changedCity
-    };
-  },
-
-  changeActiveOffers: (newOffers) => {
-    return {
-      type: ActionType.CHANGE_OFFERS,
-      payload: newOffers
-    };
-  }
 };
 
 export const reducer = (state = initialState, action) => {
